test(mitm-addon): split request hook tests by owner
