+++ crates/runner/mitm-addon/tests/test_request_handler_authority_validation.py
@@ -1,709 +1,20 @@
 """Authority validation request hook integration tests."""
 
-import ipaddress
 import json
-from collections.abc import Iterator
-from typing import cast
 
 import pytest
-from mitmproxy import connection
 
-import builtin_host_policy
 import flow_metadata_keys as metadata_keys
 import mitm_addon
-import registry
-import request_classification
 import upstream_destination_binding
-from body_limits import STREAM_BUFFER_LIMIT
 from tests.jsonl_log_helpers import read_jsonl_entries_after_flush
-from tests.registry_helpers import write_trusted_catalog_cache_text
-from tests.request_handler_helpers import (
-    _single_firewall_vm,
-    _write_github_firewall_registry,
-    _write_registry,
-)
-from tests.requestheaders_helpers import (
-    _assert_no_request_stream,
-    await_requestheaders_result,
-)
-from tests.upstream_connection_helpers import mark_connected_tls_upstream
+from tests.request_handler_helpers import _write_github_firewall_registry
+from tests.upstream_connection_helpers import bind_flow_upstream
 
 _BROWSER_USER_AGENT = (
     "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 "
     "(KHTML, like Gecko) HeadlessChrome/126.0.0.0 Safari/537.36"
 )
-
-
-class _FailOnIterationList(list[str]):
-    def __iter__(self) -> Iterator[str]:
-        raise AssertionError("resolved host policy lists must not be reparsed during requests")
-
-
-class _FailOnIterationDict(dict[str, object]):
-    def __iter__(self) -> Iterator[str]:
-        raise AssertionError("resolved host policy objects must not be reparsed during requests")
-
-
-def _write_test_oauth_registry(tmp_path):
-    return _write_registry(
-        tmp_path,
-        client_ip="10.200.0.5",
-        vm_info=_single_firewall_vm(
-            tmp_path,
-            firewall_name="test-oauth",
-            api_entry={
-                "base": "https://api.vm0.ai/api/test/oauth-provider",
-                "auth": {"headers": {"Authorization": "Bearer x"}},
-                "permissions": [{"name": "echo", "rules": ["GET /echo"]}],
-            },
-            network_policy={
-                "allow": ["echo"],
-                "deny": [],
-                "ask": [],
-                "unknownPolicy": "deny",
-            },
-        ),
-    )
-
-
-def _write_host_policy_registry(
-    tmp_path,
-    *,
-    firewall_name: str,
-    base: str,
-    host_policy: dict[str, object],
-    marked_builtin: bool = True,
-    vm_fields: dict[str, object] | None = None,
-):
-    api_entry: dict[str, object] = {
-        "base": base,
-        "auth": {"headers": {"Authorization": "Bearer ${{ secrets.TEST_TOKEN }}"}},
-        "hostPolicy": host_policy,
-        "permissions": [{"name": "full-access", "rules": ["ANY /{path+}"]}],
-    }
-    if marked_builtin:
-        api_entry[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER] = True
-    return _write_registry(
-        tmp_path,
-        client_ip="10.200.0.5",
-        vm_info=_single_firewall_vm(
-            tmp_path,
-            firewall_name=firewall_name,
-            api_entry=api_entry,
-            network_policy={
-                "allow": ["full-access"],
-                "deny": [],
-                "ask": [],
-                "unknownPolicy": "deny",
-            },
-            vm_fields=vm_fields,
-        ),
-    )
-
-
-def _write_resolved_host_policy_registry(
-    tmp_path,
-    *,
-    firewall_name: str,
-    base: str,
-    host_policy: dict[str, object],
-):
-    registry_path = _write_registry(
-        tmp_path,
-        client_ip="10.200.0.5",
-        vm_info={
-            "runId": "run-resolved-host-policy",
-            "sandboxToken": "tok-resolved-host-policy",
-            "encryptedSecrets": "iv:tag:data",
-            "networkLogPath": str(tmp_path / "net.jsonl"),
-            "proxyLogPath": str(tmp_path / "proxy.jsonl"),
-            "firewalls": [{"kind": "builtin", "name": firewall_name}],
-            "networkPolicies": {
-                firewall_name: {
-                    "allow": ["full-access"],
-                    "deny": [],
-                    "ask": [],
-                    "unknownPolicy": "deny",
-                }
-            },
-        },
-    )
-    write_trusted_catalog_cache_text(
-        tmp_path / "builtin-firewall-catalog-cache.json",
-        json.dumps(
-            {
-                "schemaVersion": 1,
-                "catalogDigest": (
-                    "sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-                ),
-                "catalogVersion": "catalog-test",
-                "updatedAt": "2026-07-07T00:00:00.000Z",
-                "firewalls": {
-                    firewall_name: {
-                        "name": firewall_name,
-                        "apis": [
-                            {
-                                "base": base,
-                                "auth": {
-                                    "headers": {"Authorization": "Bearer ${{ secrets.TEST_TOKEN }}"}
-                                },
-                                "hostPolicy": host_policy,
-                                "permissions": [
-                                    {
-                                        "name": "full-access",
-                                        "rules": ["ANY /{path+}"],
-                                    }
-                                ],
-                            }
-                        ],
-                    }
-                },
-            },
-            sort_keys=True,
-        ),
-    )
-    return registry_path
-
-
-def _bind_flow_upstream(
-    flow,
-    *,
-    host: str = "api.github.com",
-    port: int = 443,
-    kinds: frozenset[upstream_destination_binding.BindingKind] = frozenset(("connector_auth",)),
-) -> None:
-    original_address = flow.server_conn.address
-    flow.server_conn.address = (host, port)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        host=host,
-        port=port,
-        kinds=kinds,
-        original_address=original_address,
-    )
-
-
-async def test_runtime_host_policy_blocks_public_destination_private_endpoint(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://strapi.example.com",
-        host_policy={"kind": "publicDestination"},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="10.0.0.5",
-        sni="strapi.example.com",
-        path="/api/articles",
-        request_headers=headers(("Host", "strapi.example.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "unsafe_public_destination"
-    assert body["reason"] == "non_public_destination"
-    assert body["destination_host"] == "10.0.0.5"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "DENY"
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "unsafe_public_destination"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_allows_public_destination_public_endpoint(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://strapi.example.com",
-        host_policy={"kind": "publicDestination"},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="8.8.8.8",
-        sni="strapi.example.com",
-        path="/api/articles",
-        request_headers=headers(("Host", "strapi.example.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    auth_fetch.assert_awaited_once()
-
-
-async def test_runtime_host_policy_blocks_public_destination_hostname_without_ip_endpoint(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://strapi.example.com",
-        host_policy={"kind": "publicDestination"},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="strapi.example.com",
-        path="/api/articles",
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "unsafe_public_destination"
-    assert body["reason"] == "invalid_destination"
-    assert body["destination_host"] == "strapi.example.com"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "DENY"
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "unsafe_public_destination"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_blocks_public_destination_private_request_host(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://10.0.0.5",
-        host_policy={"kind": "publicDestination"},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="10.0.0.5",
-        path="/api/articles",
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "unsafe_public_destination"
-    assert body["reason"] == "non_public_destination"
-    assert body["destination_host"] == "10.0.0.5"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "DENY"
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "unsafe_public_destination"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_enforces_inline_public_destination_policy(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://strapi.example.com",
-        host_policy={"kind": "publicDestination"},
-        marked_builtin=False,
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="10.0.0.5",
-        sni="strapi.example.com",
-        path="/api/articles",
-        request_headers=headers(("Host", "strapi.example.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "unsafe_public_destination"
-    assert body["reason"] == "non_public_destination"
-    assert body["destination_host"] == "10.0.0.5"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "DENY"
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "unsafe_public_destination"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_rejects_malformed_provider_policy(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="malformed-provider",
-        base="https://api.com",
-        host_policy={"kind": "providerOwned", "suffixes": ["com"]},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.com",
-        path="/v1/items",
-        request_headers=headers(("Host", "api.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "builtin_host_policy_denied"
-    assert body["reason"] == "invalid_host_policy"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-@pytest.mark.parametrize(
-    ("base", "host_header", "sni", "port", "expected_reason"),
-    [
-        (
-            "https://attacker.example.com",
-            "attacker.example.com",
-            "attacker.example.com",
-            443,
-            "provider_host_not_allowed",
-        ),
-        (
-            "https://acme.atlassian.net:8443",
-            "acme.atlassian.net:8443",
-            "acme.atlassian.net",
-            8443,
-            "provider_non_default_port",
-        ),
-    ],
-)
-async def test_runtime_host_policy_blocks_provider_owned_request_authority(
-    tmp_path,
-    real_flow,
-    mitm_ctx,
-    fake_firewall_headers,
-    headers,
-    base,
-    host_header,
-    sni,
-    port,
-    expected_reason,
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base=base,
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        port=port,
-        sni=sni,
-        path="/rest/api/3/myself",
-        request_headers=headers(("Host", host_header)),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "builtin_host_policy_denied"
-    assert body["reason"] == expected_reason
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_allows_provider_owned_request_authority(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base="https://acme.atlassian.net",
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="acme.atlassian.net",
-        path="/rest/api/3/myself",
-        request_headers=headers(("Host", "acme.atlassian.net")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    auth_fetch.assert_awaited_once()
-
-
-async def test_runtime_host_policy_blocks_requestheaders_credential_injection(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base="https://attacker.example.com",
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-        vm_fields={"captureNetworkBodies": True},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="attacker.example.com",
-        method="POST",
-        path="/rest/api/3/myself",
-        request_headers=headers(
-            ("Host", "attacker.example.com"),
-            ("Content-Length", str(STREAM_BUFFER_LIMIT + 1)),
-        ),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers(headers={"Authorization": "Bearer resolved"}) as auth_fetch,
-    ):
-        requestheaders_result = mitm_addon.requestheaders(flow)
-        await await_requestheaders_result(requestheaders_result)
-
-        binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-        assert binding.host == "attacker.example.com"
-        assert binding.kinds == frozenset(("connector_auth",))
-        auth_fetch.assert_not_called()
-        _assert_no_request_stream(flow)
-        assert mitm_addon._FIREWALL_AUTH_APPLIED_IN_REQUESTHEADERS not in flow.metadata
-        assert "Authorization" not in flow.request.headers
-        assert flow.response is None
-
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "builtin_host_policy_denied"
-    assert body["reason"] == "provider_host_not_allowed"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_runtime_host_policy_allows_requestheaders_credential_injection(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base="https://acme.atlassian.net",
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-        vm_fields={"captureNetworkBodies": True},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="acme.atlassian.net",
-        method="POST",
-        path="/rest/api/3/myself",
-        request_headers=headers(
-            ("Host", "acme.atlassian.net"),
-            ("Content-Length", str(STREAM_BUFFER_LIMIT + 1)),
-        ),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers(headers={"Authorization": "Bearer resolved"}) as auth_fetch,
-    ):
-        requestheaders_result = mitm_addon.requestheaders(flow)
-        await await_requestheaders_result(requestheaders_result)
-
-        binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-        assert binding.host == "acme.atlassian.net"
-        assert binding.kinds == frozenset(("connector_auth",))
-        auth_fetch.assert_awaited_once()
-        assert flow.request.headers["Authorization"] == "Bearer resolved"
-        assert callable(flow.request.stream)
-        assert metadata_keys.REQUEST_STREAM_BUFFER in flow.metadata
-        assert flow.metadata[mitm_addon._FIREWALL_AUTH_APPLIED_IN_REQUESTHEADERS] is True
-        assert request_classification.REQUEST_CLASSIFICATION_METADATA_KEY in flow.metadata
-
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert request_classification.REQUEST_CLASSIFICATION_METADATA_KEY not in flow.metadata
-    assert flow.metadata[metadata_keys.REQUEST_STREAM_COMPLETE] is True
-
-
-async def test_resolved_host_policy_reuses_compiled_policy_across_requests(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_resolved_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base="https://acme.atlassian.net",
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-    )
-    flows = [
-        real_flow(
-            with_response=False,
-            client_ip="10.200.0.5",
-            host="203.0.113.10",
-            sni="acme.atlassian.net",
-            path=f"/rest/api/3/items/{index}",
-            request_headers=headers(("Host", "acme.atlassian.net")),
-        )
-        for index in range(2)
-    ]
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        context = registry.get_vm_context("10.200.0.5", str(reg_path))
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        api = vm_info["firewalls"][0]["apis"][0]
-        assert isinstance(
-            api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER],
-            builtin_host_policy.CompiledBuiltinHostPolicy,
-        )
-        raw_host_policy = api["hostPolicy"]
-        assert isinstance(raw_host_policy, dict)
-        assert raw_host_policy["suffixes"] == ["atlassian.net"]
-        raw_host_policy["suffixes"] = _FailOnIterationList(["atlassian.net"])
-
-        for flow in flows:
-            await mitm_addon.request(flow)
-
-    for flow in flows:
-        assert flow.response is None
-        assert flow.request.headers["Authorization"] == "Bearer x"
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    assert auth_fetch.await_count == 2
-
-
-async def test_resolved_public_destination_host_policy_uses_compiled_policy(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_resolved_host_policy_registry(
-        tmp_path,
-        firewall_name="strapi",
-        base="https://strapi.example.com",
-        host_policy={"kind": "publicDestination"},
-    )
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="8.8.8.8",
-        sni="strapi.example.com",
-        path="/api/articles",
-        request_headers=headers(("Host", "strapi.example.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        context = registry.get_vm_context("10.200.0.5", str(reg_path))
-        assert context is not None
-        vm_info, compiled_firewalls, _ = context
-        assert compiled_firewalls is not None
-        api = vm_info["firewalls"][0]["apis"][0]
-        assert isinstance(
-            api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER],
-            builtin_host_policy.CompiledBuiltinHostPolicy,
-        )
-        raw_host_policy = api["hostPolicy"]
-        assert raw_host_policy == {"kind": "publicDestination"}
-        api["hostPolicy"] = _FailOnIterationDict({"kind": "publicDestination"})
-
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    auth_fetch.assert_awaited_once()
-
-
-async def test_runtime_host_policy_rejects_invalid_runtime_marker(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_host_policy_registry(
-        tmp_path,
-        firewall_name="jira",
-        base="https://acme.atlassian.net",
-        host_policy={"kind": "providerOwned", "suffixes": ["atlassian.net"]},
-    )
-    registry_data = json.loads(reg_path.read_text())
-    api = registry_data["vms"]["10.200.0.5"]["firewalls"][0]["firewall"]["apis"][0]
-    api[builtin_host_policy.BUILTIN_HOST_POLICY_RUNTIME_MARKER] = "invalid"
-    reg_path.write_text(json.dumps(registry_data))
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="acme.atlassian.net",
-        path="/rest/api/3/myself",
-        request_headers=headers(("Host", "acme.atlassian.net")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "builtin_host_policy_denied"
-    assert body["reason"] == "invalid_host_policy"
-    auth_fetch.assert_not_called()
-    assert "Authorization" not in flow.request.headers
 
 
 @pytest.mark.parametrize(
@@ -841,532 +152,6 @@ async def test_browser_user_agent_marker_survives_authority_validation_block(
     assert entry["browser_user_agent"] is True
 
 
-async def test_matching_sni_and_host_blocks_firewall_auth_when_upstream_is_unbound(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "upstream_destination_unbound"
-    assert body["reason"] == "connector_auth"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-    with mitm_ctx():
-        mitm_addon.response(flow)
-
-    [entry] = read_jsonl_entries_after_flush(tmp_path / "net.jsonl")
-    assert entry["action"] == "BLOCK"
-    assert entry["firewall_error"] == "upstream_destination_unbound"
-    assert entry["upstream_binding_reason"] == "connector_auth"
-    assert entry["upstream_binding_trusted_host"] == "api.github.com"
-    assert entry["upstream_binding_request_host"] == "203.0.113.10"
-    assert entry["upstream_binding_request_port"] == 443
-    assert entry["upstream_binding_server_connected"] is True
-    assert entry["upstream_binding_server_address"] == "203.0.113.10:443"
-    assert entry["upstream_binding_direct_binding_present"] is False
-    assert entry["upstream_binding_client_binding_count"] == 0
-
-
-async def test_matching_sni_and_host_retargets_unconnected_firewall_auth(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers(),
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.server_conn.address == ("api.github.com", 443)
-    assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.github.com"
-    assert binding.kinds == frozenset(("connector_auth",))
-    assert binding.original_address == ("203.0.113.10", 443)
-
-
-async def test_matching_sni_and_host_records_binding_when_unconnected_address_already_matches(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="api.github.com",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers(),
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.server_conn.address == ("api.github.com", 443)
-    assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.github.com"
-    assert binding.kinds == frozenset(("connector_auth",))
-    assert binding.original_address == ("api.github.com", 443)
-
-
-async def test_matching_sni_and_host_blocks_connected_firewall_auth_without_verified_tls(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="140.82.112.5",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("140.82.112.5", 443)
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_blocks_connected_firewall_auth_when_upstream_sni_differs(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="140.82.112.5",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("140.82.112.5", 443)
-    flow.server_conn.sni = "attacker.example.com"
-    flow.server_conn.timestamp_tls_setup = 1.0
-    flow.server_conn.certificate_list = (object(),)
-    flow.server_conn.error = None
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_blocks_connected_firewall_auth_when_upstream_tls_failed(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="140.82.112.5",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("140.82.112.5", 443)
-    flow.server_conn.sni = "api.github.com"
-    flow.server_conn.timestamp_tls_setup = 1.0
-    flow.server_conn.certificate_list = (object(),)
-    flow.server_conn.error = "certificate verify failed"
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_allows_connected_firewall_auth_with_early_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="140.82.112.5",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("140.82.112.5", 443)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("140.82.112.5", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.server_conn.address == ("140.82.112.5", 443)
-    assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-    assert flow.request.headers["Authorization"] == "Bearer x"
-
-
-async def test_matching_sni_and_host_allows_connected_firewall_auth_after_retargeting(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.github.com",
-        server_address=("api.github.com", 443),
-        peername=("140.82.112.5", 443),
-    )
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("203.0.113.10", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.server_conn.address == ("api.github.com", 443)
-    assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.original_address == ("140.82.112.5", 443)
-
-
-async def test_matching_sni_and_host_allows_connected_firewall_auth_with_verified_tls_no_peername(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.github.com",
-        server_address=("api.github.com", 443),
-        peername=None,
-        client_sockname=("140.82.112.5", 443),
-    )
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("203.0.113.10", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.original_address == ("140.82.112.5", 443)
-
-
-async def test_matching_sni_and_host_blocks_connected_firewall_auth_without_endpoint_proof(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.address = ("api.github.com", 443)
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = None
-    flow.client_conn.sockname = ("127.0.0.1", 8080)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("203.0.113.10", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-@pytest.mark.parametrize(
-    "client_sockname",
-    [
-        pytest.param(cast(tuple[str, int], ("203.0.113.10",)), id="short-tuple"),
-        pytest.param(
-            cast(
-                tuple[str, int],
-                (int(ipaddress.IPv4Address("203.0.113.10")), 443),
-            ),
-            id="integer-host",
-        ),
-        pytest.param(("127.0.0.1", 443), id="loopback"),
-        pytest.param((str(ipaddress.IPv4Address(0)), 443), id="unspecified-ipv4"),
-        pytest.param(("::", 443), id="unspecified-ipv6"),
-    ],
-)
-async def test_matching_sni_and_host_blocks_non_authoritative_connected_endpoint(
-    tmp_path,
-    real_flow,
-    mitm_ctx,
-    fake_firewall_headers,
-    headers,
-    client_sockname: tuple[str, int],
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.github.com",
-        server_address=("api.github.com", 443),
-        peername=None,
-        client_sockname=client_sockname,
-    )
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("203.0.113.10", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_allows_equivalent_ipv6_binding_peer(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.github.com",
-        server_address=("api.github.com", 443),
-        peername=None,
-    )
-    flow.server_conn.peername = cast(
-        tuple[str, int],
-        ("2001:4860:4860:0:0:0:0:8888", 443, 0, 0),
-    )
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("2001:4860:4860::8888", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-
-
-async def test_matching_sni_and_host_blocks_connected_firewall_auth_with_stale_binding_peer(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.99",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("203.0.113.99", 443)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("connector_auth",)),
-        original_address=("140.82.112.5", 443),
-    )
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_allows_bound_firewall_auth(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers
-):
-    reg_path = _write_github_firewall_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.github.com",
-        path="/repos",
-        request_headers=headers(("Host", "api.github.com")),
-    )
-    _bind_flow_upstream(flow)
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers(),
-    ):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-    assert flow.metadata[metadata_keys.FIREWALL_NAME] == "github"
-    assert flow.metadata[metadata_keys.FIREWALL_PERMISSION] == "full-access"
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    assert flow.metadata[metadata_keys.ORIGINAL_URL] == "https://api.github.com/repos"
-    assert flow.metadata[metadata_keys.NETWORK_LOG_TARGET] == {
-        "url": "https://api.github.com/repos",
-        "host": "api.github.com",
-        "port": 443,
-    }
-
-
 @pytest.mark.parametrize("http_version", ["HTTP/2.0", "HTTP/3"])
 @pytest.mark.parametrize("regular_host", [None, "api.github.com"], ids=["authority-only", "host"])
 async def test_valid_pseudo_authority_allows_firewall_auth(
@@ -1390,7 +175,7 @@ async def test_valid_pseudo_authority_allows_firewall_auth(
     )
     flow.request.http_version = http_version
     flow.request.authority = "api.github.com"
-    _bind_flow_upstream(flow)
+    bind_flow_upstream(flow)
 
     with (
         mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
@@ -1480,411 +265,6 @@ async def test_rejects_spoofed_host_before_vm0_api_auto_allow(
     body = json.loads(flow.response.content)
     assert body["error"] == "authority_mismatch"
     assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "DENY"
-
-
-async def test_matching_sni_and_host_blocks_connected_vm0_api_edge_when_unbound(
-    registry_file, real_flow, mitm_ctx, headers
-):
-    flow = real_flow(
-        with_response=False,
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/webhooks/agent/heartbeat",
-        request_headers=headers(("Host", "api.vm0.ai")),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-
-    with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-        await mitm_addon.request(flow)
-
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    body = json.loads(flow.response.content)
-    assert body["error"] == "upstream_destination_unbound"
-    assert body["reason"] == "api_allow"
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-
-
-async def test_matching_sni_and_host_allows_authenticated_connected_vm0_api_edge(
-    registry_file, real_flow, mitm_ctx, headers
-):
-    flow = real_flow(
-        with_response=False,
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        method="POST",
-        path="/api/webhooks/agent/heartbeat",
-        request_headers=headers(
-            ("Authorization", "Bearer tok-xyz"),
-            ("Host", "api.vm0.ai"),
-        ),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.vm0.ai",
-        server_address=("203.0.113.10", 443),
-        peername=("203.0.113.10", 443),
-    )
-
-    with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.vm0.ai"
-    assert binding.kinds == frozenset(("api_allow",))
-    assert binding.original_address == ("203.0.113.10", 443)
-
-
-async def test_matching_sni_and_host_allows_test_connector_on_authenticated_api_edge(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "preview-secret"),
-        ),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.vm0.ai",
-        server_address=("203.0.113.10", 443),
-        peername=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.vm0.ai"
-    assert binding.kinds == frozenset(("connector_auth",))
-    assert binding.original_address == ("203.0.113.10", 443)
-
-
-async def test_test_connector_extends_existing_api_allow_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "preview-secret"),
-        ),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("203.0.113.10", 443)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.vm0.ai",
-        port=443,
-        kinds=frozenset(("api_allow",)),
-        original_address=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_awaited_once()
-    assert flow.response is None
-    assert flow.request.headers["Authorization"] == "Bearer x"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.vm0.ai"
-    assert binding.kinds == frozenset(("api_allow", "connector_auth"))
-    assert binding.original_address == ("203.0.113.10", 443)
-
-
-async def test_test_connector_without_bypass_does_not_extend_existing_api_allow_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "wrong-secret"),
-        ),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.vm0.ai",
-        port=443,
-        kinds=frozenset(("api_allow",)),
-        original_address=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-    assert upstream_destination_binding.binding_snapshot_for_tests() == {}
-
-
-async def test_test_connector_unconnected_without_bypass_blocks_before_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "wrong-secret"),
-        ),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-    assert upstream_destination_binding.binding_snapshot_for_tests() == {}
-
-
-async def test_test_connector_without_bypass_does_not_reuse_connector_auth_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "wrong-secret"),
-        ),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    flow.server_conn.peername = ("203.0.113.10", 443)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.vm0.ai",
-        port=443,
-        kinds=frozenset(("api_allow", "connector_auth")),
-        original_address=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-    assert upstream_destination_binding.binding_snapshot_for_tests() == {}
-
-
-async def test_test_connector_rejects_stale_unconnected_api_allow_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.99",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "preview-secret"),
-        ),
-    )
-    flow.server_conn.address = ("203.0.113.99", 443)
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.vm0.ai",
-        port=443,
-        kinds=frozenset(("api_allow",)),
-        original_address=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-    assert upstream_destination_binding.binding_snapshot_for_tests() == {}
-
-
-async def test_test_connector_rejects_mismatched_existing_binding(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "preview-secret"),
-        ),
-    )
-    flow.server_conn.state = connection.ConnectionState.OPEN
-    upstream_destination_binding.record_server_binding(
-        flow.server_conn,
-        client=flow.client_conn,
-        host="api.github.com",
-        port=443,
-        kinds=frozenset(("api_allow",)),
-        original_address=("203.0.113.10", 443),
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-    assert upstream_destination_binding.binding_snapshot_for_tests() == {}
-
-
-async def test_matching_sni_and_host_blocks_test_connector_api_edge_without_bypass(
-    tmp_path, real_flow, mitm_ctx, fake_firewall_headers, headers, monkeypatch
-):
-    reg_path = _write_test_oauth_registry(tmp_path)
-    flow = real_flow(
-        with_response=False,
-        client_ip="10.200.0.5",
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/test/oauth-provider/echo",
-        request_headers=headers(
-            ("Host", "api.vm0.ai"),
-            ("x-vm0-test-endpoint-bypass", "wrong-secret"),
-        ),
-    )
-    mark_connected_tls_upstream(
-        flow,
-        sni="api.vm0.ai",
-        server_address=("203.0.113.10", 443),
-        peername=None,
-    )
-    monkeypatch.setenv("VERCEL_AUTOMATION_BYPASS_SECRET", "preview-secret")
-
-    with (
-        mitm_ctx(registry_path=str(reg_path), api_url="https://api.vm0.ai"),
-        fake_firewall_headers() as auth_fetch,
-    ):
-        await mitm_addon.request(flow)
-
-    auth_fetch.assert_not_called()
-    assert flow.response is not None
-    assert flow.response.status_code == 403
-    assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "upstream_destination_unbound"
-    assert "Authorization" not in flow.request.headers
-
-
-async def test_matching_sni_and_host_retargets_unconnected_vm0_api_auto_allow(
-    registry_file, real_flow, mitm_ctx, headers
-):
-    flow = real_flow(
-        with_response=False,
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/runs/heartbeat",
-        request_headers=headers(("Host", "api.vm0.ai")),
-    )
-
-    with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.server_conn.address == ("api.vm0.ai", 443)
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-    binding = upstream_destination_binding.binding_snapshot_for_tests()[flow.server_conn.id]
-    assert binding.host == "api.vm0.ai"
-    assert binding.kinds == frozenset(("api_allow",))
-    assert binding.original_address == ("203.0.113.10", 443)
-
-
-async def test_matching_sni_and_host_allows_bound_vm0_api_auto_allow(
-    registry_file, real_flow, mitm_ctx, headers
-):
-    flow = real_flow(
-        with_response=False,
-        host="203.0.113.10",
-        sni="api.vm0.ai",
-        path="/api/runs/heartbeat",
-        request_headers=headers(("Host", "api.vm0.ai")),
-    )
-    _bind_flow_upstream(flow, host="api.vm0.ai", kinds=frozenset(("api_allow",)))
-
-    with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-        await mitm_addon.request(flow)
-
-    assert flow.response is None
-    assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
 
 
 @pytest.mark.parametrize(
